requirements.txt
apache-airflow==2.7.3
pyspark==3.5.1
pandas==2.2.2
requests==2.31.0 
Flask-Session==0.3.2


